requirements.dev.txt
-r requirements.txt
pyyaml
pylama
tox
mock
pytest
pytest-cov
coverage
coveralls
